feat: Improve version handling for OQMD

+++ src/lematerial_fetcher/fetcher/oqmd/utils.py
@@ -43,13 +43,14 @@ def download_and_process_oqmd_sql(
         version_db_config = db_config.copy()
         version_db_config["database"] = f"{db_config['database']}_version"
         version_db = MySQLDatabase(**version_db_config)
+
+        version_db.create_database()
         current_url, current_modification_date = get_oqmd_version_if_exists(version_db)
 
         if current_url == latest_url and current_modification_date == modification_date:
             logger.info(
                 "OQMD database is already at the latest version. Skipping download."
             )
-            update_oqmd_version(version_db, db_config, latest_url, modification_date)
             return
 
         logger.info("New OQMD version detected. Proceeding with download...")
@@ -77,6 +78,15 @@ def download_and_process_oqmd_sql(
 
         # Create fresh database
         logger.info("Setting up MySQL database")
+
+        user_input = input(
+            f"Warning: This will drop the existing database to replace it with the new one from {latest_url}. "
+            "Are you sure you want to continue? If you press N, the script will execute with your current database. (y/N): "
+        )
+        if user_input.lower() != "y":
+            logger.info("Database update cancelled by user")
+            return
+
         db.drop_database()
         db.create_database()
 
@@ -152,12 +162,12 @@ def get_oqmd_version_if_exists(
     """)
 
     result = version_db.fetch_items(
-        f"SELECT download_url, last_updated FROM {version_db_name} ORDER BY last_updated DESC LIMIT 1"
+        query=f"SELECT download_url, last_updated FROM {version_db_name} ORDER BY last_updated DESC LIMIT 1"
     )
     version_db.close()
 
     if not result:
-        return None
+        return None, None
     return result[0]["download_url"], result[0]["last_updated"]
